refactor(analysis-types): split license DTOs
